requirements.txt
spotipy
gitpython